Fix chunk uploads being attributed to wrong world
Previously any chunk uploads were always attributed to the currently
active world. If such an upload is scheduled by one of the render
worker, it belongs to a different world thread (the one the render
worker had been assigned to).
This caused those uploads to be stored in the wrong world state while
the render cache builder allocations were stored in the correct one.
If the incorrect world was subsequently unloaded, those chunks and
render cache builder were locked away which caused those chunks to not
be rendered until they are updated.
Additionally this caused a live lock in stopChunkUpdates when exiting
the main world.