Fix incorrect alpha blending state after portal is rendered (fixes #53)
The workaround fixed a problem that wasn't actually really there (well,
it kind of was, the alpha state was indeed in an inconsistent state,
it should have always been enabled).
Anyhow, the proper way to fix our issue (no rendering of the inactive
portal) is to disable TEXTURE_2D while we render it, that way we get
the color we specified (previously we were just lucky and got a random,
completely black but also completely transparent texture).