Fix camera entity position when player is in different view (fixes #54)
We currently update camera entities before each render pass based on the
one in the parent render pass. This works while the root pass is the one
containing the player but when it doesn't, we need to update the root
pass based on the view containing the player before rendering.
I.e. we've previously distributed the camera position downwards from the
player's main view but we also need to distribute it upwards.

This becomes less of an issue once we stop rendering purely based on
the view entity positions and start using the computed render pass
camera positions instead (probably v0.2).