Fix using the same portal on two consecutive ticks (fixes #67)
The important thing here are the calls to the remote portal's
checkTeleportees (which updates the tracked position for the entity).

Removing thisTickPos just makes a call to checkTeleportees sufficient
for doing that.

Ticking all client worlds is necessary as otherwise the remote portal
will still have an old prevTickPos for the player even though they long
left and therefore immediately teleport them back.