Fix incorrect main view when using a portal with camera on original side
Given we use calculatePlaneIntercept, we no longer need to check for
the parent portal to prevent going back.
In fact, we must not check for the parent portal as going back must be
allowed e.g. when we start on side A and due to time interpolation move
to side B and then due to actual camera position have to move back to A.