Stop moving around ViewEntity and use dedicated camera entity instead
The camera entity does not update the ViewFrustum position.

This should fix any bugs where terrain in portals becomes invisible
when another portal is loaded.

What remains is that portals which are further from the view entity
than the client render distance won't show much terrain (not an issue
if it is greater than the server view distance).
We probably want a RenderPass-aware ViewFrustum to fix that (and in that
case must do occlusion culling only after tree building, never during).

This also changes the max distance for view sharing to half the view
distance instead of fixed 64 blocks.