Re-write client-side view swapping (#151)
Primarily because we need the player in the old view during packet
handling and there may already be a player entity in the new one.
See [ClientView.clientPlayer].

Note: This is not sufficient to fix the aforementioned issue but it's
important anyway, e.g. if the player's inventory is updated after
client-side swap but before server-side swap, then this now correctly
updates the inventory on the client as well. Previously the view
entity's inventory would be updated.