Fix some entities sometimes not syncing when new view is added
Happened when using a telepad (without enhanced third-party transfers)
which is in the same chunk as a nether portal to teleport into the
nether to a telepad which is in the same chunk as the remote to the
nether portal. Under those circumstances, the overworld nether portal
entity will not immediately be synced (only after some delay).