Fix Forge's sky color cache to work with multiple worlds (fixes #323)
To reproduce the original issue, one must ensure that the cache
from the two worlds collides, i.e. the x and z block coordinates of the
camera must be the same in both worlds.
Given the Aether is largely empty, there's a good chance a portal will
generate at the same coordinates over there as it was in the overworld.