Support worlds where provider ID is not predetermined from dimension ID (fixes #383)
When hot-loading worlds on a sponge server (e.g. with ProjectWorld),
those worlds can have e.g. dim ID 2 but use the vanilla overworld
provider.
The Forge client determines the provider to use from the dim ID. When
teleporting normally, Sponge manually registers the new provider mapping
(see [1]). This however does not happen when BP loads the world.

This commit simply adds the provider name the the CreateWorld message
and registers it to the dim ID before attempting to create a world.

[1]: https://github.com/SpongePowered/SpongeForge/blob/3fbc8850d9a5a31fecc5bd33693a42c1e1396e34/src/main/java/org/spongepowered/mod/mixin/core/common/world/WorldManagerMixin_Forge.java#L106-L124