Disambiguate generic canRenderName mixin target
While mixin recognizes that one of the two possibilities is a synthetic
method, that information is not available to remap and so it complains
about the ambiguity.
Additionally there's apparently no ambiguity when mapping from fabric to
forge, so we need to manually specify that as well (otherwise it'll
produce ambiguous 1.14.4-forge targets which will error when converting
to 1.12.2).