Skip view ticking when disconnected (fixes #432)
There's a small (but significant) window of time between when the
Disconnect button is pressed (and mc.playerController is set to null)
and when the channel is closed (where forge fires the disconnect event
and we unregister all remaining views).
If we tick views during that window of time, the client will crash, so
we now just no longer do that (doesn't serve any purpose anyway).