Fix third-party transfer to world with active portal (fixes #455)
See the comment in beforeTransferToDimension for how vanilla is broken.

Previously BP replaced the old WorldManager with the new one when the
player is added to the new world's PlayerChunkMap (which is the last
step of vanilla's transfer procedure) but that's too late if new views
are registered when the player is added to the world (e.g. if there are
active portals at the destination).
Now we do it right before the respawn packet (the first step of
vanilla's transfer procedure).