Switch from registry overwrites to mixins for vanilla integration
Registry overwrites are a Forge thing, so we cannot use it with Fabric.
Side note: you can overwrites entries in the vanilla registry but any
class which had already obtained a references by then will continue
using the old one (Forge replaced all the references with @ObjectHolder
magic, hence why it worked fine there). With that in mind, using mixins
also feels a lot cleaner of an approach.