Split SourceSets into separate mods and bundle them via jar-in-jar
This allows the user to pick and choose which parts of the mod they want
in a rather clean way (just not install that module).
And for all those that don't case, they can just install the bundle jar.

Well, except for 1.14 Forge because there jar-in-jar is broken:
https://github.com/MinecraftForge/MinecraftForge/issues/6239

Note that I intend to split the vanilla source set into a nether and an
end mod and have already done so for forge but not yet for fabric,
that'll happen in one of the following commits (together with splitting
the source set itself).

Also gets 1.14 Forge running outside the dev env but not quite out of
the box:
- Must manually install Kottle because jar-in-jar is broken
- Must manually install MixinBootstrap
- Must manually upgrade eventbus to 1.1.0 for SpongePowered/Mixin#369